enqueue_* now take float seconds.

+++ server/tasks.h
@@ -37,9 +37,9 @@ extern int tasks_set_connection_option(task_queue, const char *,
 
 extern void new_input_task(task_queue, const char *);
 extern enum error enqueue_forked_task2(activation a, int f_index,
-			       unsigned after_seconds, int vid);
+			       double after_seconds, int vid);
 extern enum error enqueue_suspended_task(vm the_vm, void *data);
-				/* data == &(int after_seconds) */
+				/* data == &(double after_seconds) */
 extern enum error make_reading_task(vm the_vm, void *data);
 				/* data == &(Objid connection) */
 extern void resume_task(vm the_vm, Var value);
@@ -110,6 +110,9 @@ extern db_verb_handle find_verb_for_programming(Objid player,
 
 /* 
  * $Log: not supported by cvs2svn $
+ * Revision 1.3  1998/12/14 13:19:08  nop
+ * Merge UNSAFE_OPTS (ref fixups); fix Log tag placement to fit CVS whims
+ *
  * Revision 1.2.2.1  1998/12/06 07:13:23  bjj
  * Rationalize enqueue_forked_task interface and fix program_ref leak in
  * the case where fork fails with E_QUOTA.  Make .queued_task_limit=0 really
